Drop top-level __init__.py
Turns out this just complicates git-submodule redistribution
because for dependant projects then setuptools' find_modules()
function finds also build_manpages files and subsequent calls to
the `python setup.py install` also installs the files.  And this
is not desired.

TODO: find a way how to distribute and not install the files
automatically