Several new configuration/cmd-line options
The new 'manual_section'/--manual-section allows us to specify a desired
manual page version.

The new 'manual_title'/--manual--title option allows us to specify
manual page title.

The 'version'/--version has a better documentation now, and is fixed so
it has the effect.

The 'description'/--description now has the needed effect, and better
documentation.

The 'long_description' is deprecated, unused and hidden.  It doesn't
make sense to mixup project-description with the utility description.
Utility description is directly in the Argparse object.

Added test-cases for all the variants.

Fixes: #15
Fixes: #62