build_manpages: simplify use of setup.py command wrappers
Most of our users will not do any own command overrides to those
commands.  So it is not necessary to instruct them to import
'setuptools.command' since we can do this automatically.  Still, if
users override something themselves — we can take their overridden
commands as the argument of the particular wrapper (and also we need to
stay backward compatible).

We are slowly moving form distutils to setuptools.  We prefer using
setuptools on modern python, and if importing setuptools fails (older
python or older setuptools) we can try importing distutils as fallback.